+++ package.json
@@ -4,13 +4,14 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "mocha --timeout 50000 --recursive test"
+    "test": "mocha --reporter mochawesome --timeout 50000 --recursive test"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
     "chai": "^4.1.2",
     "mocha": "^5.2.0",
+    "mochawesome": "^3.0.3",
     "puppeteer": "^1.7.0"
   }
 }
